pyproject.toml
[project]
name = "abdownloader"
version = "0.1.0"
description = ""
authors = [
    {name = "Gal Somovic",email = "galsomovic@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-fasthtml>=0.12.9"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
